# .futurxlab/inknowing-api-spec.yaml
openapi: 3.0.3
info:
  title: InKnowing API
  description: |
    知应(InKnowing) AI知识对话平台 MVP API规范
    基于业务逻辑守恒原理设计，确保与用户旅程、时序图、状态图完全对应
  version: 1.0.0
  contact:
    name: FuturX Lab
    email: support@inknowing.com
servers:
  - url: http://localhost:3000/api
    description: 本地开发环境
  - url: https://api.inknowing.com
    description: 生产环境
...
paths:
...
  /admin/books/identify:
    post:
      tags: [Admin]
      summary: AI识别书籍
      description: |
        通过书名让AI自动识别并生成书籍信息
        状态转换: BOOK_MANAGEMENT → BOOK_IDENTIFYING → BOOK_IDENTIFIED
        引用: #REF-BOOK-IDENTIFY
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                  description: 书籍名称
      responses:
        '200':
          description: 识别结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  book_info:
                    $ref: '#/components/schemas/BookInfo'
                  ai_score:
                    type: integer
                    minimum: 1
                    maximum: 10
                    description: AI了解程度评分
                  cover_options:
                    type: array
                    items:
                      type: string
                      format: uri
...
  /admin/books/{id}/documents:
    post:
      tags: [Admin]
      summary: 上传书籍文档
      description: |
        上传TXT或Markdown文档作为主文档或补充材料
        状态转换: DOCUMENT_REQUIRED/OPTIONAL → DOCUMENT_UPLOADING → DOCUMENT_PARSING
        引用: #REF-DOC-UPLOAD
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, type]
              properties:
                file:
                  type: string
                  format: binary
                  description: TXT或Markdown文件
                type:
                  type: string
                  enum: [main, supplement]
                  description: 文档类型
      responses:
        '200':
          description: 上传成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  document_id:
                    type: string
                  status:
                    type: string
  /admin/books/{id}/vectorize:
    post:
      tags: [Admin]
      summary: 向量化处理
      description: |
        将文档内容向量化并存入ChromaDB
        状态转换: DOCUMENT_CHUNKING → VECTORIZING → VECTOR_COMPLETE
        引用: #REF-DOC-VECTOR, #REF-VECTOR-001
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                document_ids:
                  type: array
                  items:
                    type: string
                  description: 要向量化的文档ID列表
      responses:
        '200':
          description: 向量化开始
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                  status:
                    type: string
                    enum: [processing, completed, failed]
  /admin/books/{id}/status:
    put:
      tags: [Admin]
      summary: 更新书籍状态
      description: |
        上架或下架书籍
        状态转换: BOOK_STATUS_CONTROL → BOOK_ONLINE/OFFLINE
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [online, offline]
      responses:
        '200':
          description: 状态更新成功
  /admin/books/{id}:
    put:
      tags: [Admin]
      summary: 编辑书籍信息
      description: |
        更新书籍的所有信息
        状态: BOOK_MANAGEMENT
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookUpdateRequest'
      responses:
        '200':
          description: 更新成功

    delete:
      tags: [Admin]
      summary: 删除书籍
      description: |
        删除书籍及其所有相关数据
        状态: BOOK_MANAGEMENT
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 删除成功
...
components:
  schemas:
...
    BookInfo:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        description:
          type: string
        publisher:
          type: string
        publish_date:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
...
    BookUpdateRequest:
      $ref: '#/components/schemas/BookInfo'
...
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: 用户Session认证

    adminAuth:
      type: apiKey
      in: cookie
      name: admin_session
      description: 管理员Session认证

# 业务逻辑守恒验证标记
# 所有API端点都与以下视图完全对应：
# 1. User Journey: 用户旅程中的每个动作都有对应API
# 2. Sequence Diagram: 时序图中的每个交互都有API支撑
# 3. State Diagram: 状态转换的每个触发点都是API调用
# 4. 双向可推导: 从API可以还原完整业务流程